Reorganized Files & Created Autonomous1Ball
Lots of stuff... renamed things and created the autonomous file that will hold one of the competition autonomous options.
