--- src/main/java/frc/robot/commands/MotionMagicSequence.java
@@ -1,21 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.robot.commands;
-
-import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import edu.wpi.first.wpilibj2.command.WaitCommand;
-import frc.robot.subsystems.Drivetrain;
-
-// NOTE:  Consider using this command inline, rather than writing a subclass.  For more
-// information, see:
-// https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class MotionMagicSequence extends SequentialCommandGroup {
-  /** Creates a new MotionMagicSequence. */
-  public MotionMagicSequence(Drivetrain drivetrain, double distance) {
-    // Add your commands in the addCommands() call, e.g.
-    // addCommands(new FooCommand(), new BarCommand());
-    addCommands(new MotionMagicAuto(drivetrain, distance), new WaitCommand(1), new MotionMagicAuto(drivetrain, -distance));
-  }
-}
